Simplify and fix error type

+++ src/freenet/node/fcp/UploadDirRequestStatus.java
@@ -7,7 +7,7 @@ public class UploadDirRequestStatus extends UploadRequestStatus {
 	public UploadDirRequestStatus(String identifier, short persistence, boolean started, boolean finished, 
 			boolean success, int total, int min, int fetched, int fatal, int failed,
 			boolean totalFinalized, long last, short prio, // all these passed to parent
-			FreenetURI finalURI, FreenetURI targetURI, short failureCode,
+			FreenetURI finalURI, FreenetURI targetURI, int failureCode,
 			String failureReasonShort, String failureReasonLong, long size, int files) {
 		super(identifier, persistence, started, finished, success, total, min, fetched, 
 				fatal, failed, totalFinalized, last, prio, finalURI, targetURI,
